+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## 0.5.5
 
-- shell widget DSL for setAppName in SWT Display
-- shell widget DSL for setAppVersion in SWT Display
+- shell widget args for SWT Display app name and app version
 - Glimmer DSL colors lazy initialize and don't have an SWT Display object dependency anymore
 - Glimmer DSL Menu/MenuItem support